+++ src/routes/GenericListing.tsx
@@ -1,7 +1,9 @@
-import { Divider, Heading, HStack, Spacer, Box } from "@chakra-ui/react";
-import { useCallback, useEffect, useState } from "react";
-import { createSearchParams, useLoaderData, useNavigate, useSearchParams } from "react-router-dom";
+import { Divider, Heading, HStack, Spacer, Box, Select, Text } from "@chakra-ui/react";
+import { useCallback, useEffect } from "react";
+import { useLoaderData, useNavigate } from "react-router-dom";
 import { Pagination, DebouncedInput, Table } from "@diamondlightsource/ui-components";
+import { usePaginationSearchParams } from "utils/hooks";
+import { Option, Options } from "components/form/input";
 
 export interface TableProps {
   headers: {
@@ -9,6 +11,8 @@ export interface TableProps {
     label: string;
   }[];
   heading: string;
+  /** Valid sorting values */
+  sortOptions?: Option[];
   makePathCallback?: (item: Record<string, string | number>, index: number) => string;
 }
 
@@ -18,35 +22,12 @@ interface TableData {
   limit: number;
 }
 
-const GenericListing = ({ headers, heading, makePathCallback }: TableProps) => {
+const GenericListing = ({ headers, heading, sortOptions, makePathCallback }: TableProps) => {
   const data = useLoaderData() as TableData;
-  const [searchParams] = useSearchParams();
-  const [page, setPage] = useState(parseInt(searchParams.get("page") || "1"));
-  const [itemsPerPage, setItemsPerPage] = useState(parseInt(searchParams.get("items") || "20"));
-  const [search, setSearch] = useState(searchParams.get("search") || "");
+  const { page, sortBy, setPage, setItemsPerPage, setSortBy, onSearch } =
+    usePaginationSearchParams();
 
   const navigate = useNavigate();
-
-  const handleSearch = useCallback((search: string) => {
-    setPage(1);
-    setSearch(search);
-  }, []);
-
-  useEffect(
-    () =>
-      navigate(
-        {
-          pathname: ".",
-          search: `?${createSearchParams({
-            search: search,
-            page: page.toString(),
-            items: itemsPerPage.toString(),
-          })}`,
-        },
-        { replace: true }
-      ),
-    [search, page, itemsPerPage, navigate]
-  );
 
   const handleRowClicked = useCallback(
     (item: Record<string, any>, index: number) => {
@@ -66,10 +47,24 @@ const GenericListing = ({ headers, heading, makePathCallback }: TableProps) => {
       <HStack>
         <Heading>{heading}</Heading>
         <Spacer />
+        {sortOptions && (
+          <>
+            <Text id='sort-by-label'>Sort By</Text>
+            <Select
+              defaultValue={sortBy ?? 0}
+              aria-labelledby='sort-by-label'
+              variant='hi-contrast'
+              size='sm'
+              w={{ base: "auto", md: "20%" }}
+              onChange={(e) => setSortBy(e.target.value)}
+            >
+              <Options values={sortOptions} />
+            </Select>
+          </>
+        )}
         <DebouncedInput
-          onChangeEnd={handleSearch}
-          borderColor='gray.600'
-          bg='diamond.50'
+          onChangeEnd={onSearch}
+          variant='hi-contrast'
           w={{ base: "auto", md: "20%" }}
           size='sm'
           placeholder='Search...'